+++ documents/prototypage/src/Ignored.java
@@ -24,7 +24,7 @@ public class Ignored extends Exp{
 	}
 
     @Override
-    public float evaluate() throws Exception {
+    public double evaluate() throws Exception {
         throw new UnsupportedOperationException();
     }
